@@ -1,16 +1,15 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { Davit } from "./pages/Davit";
 import * as serviceWorker from "./serviceWorker";
 import { store } from "./store";
 import "./style/index.css";
 
-ReactDOM.render(
-    <Provider store={store}>
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+ root.render(<Provider store={store}>
         <Davit />
-    </Provider>,
-    document.getElementById("root"),
+    </Provider>
 );
 
 // If you want your app to work offline and load faster, you can change
