@@ -4,16 +4,16 @@ import { DavitTableRowData } from "../../../../../components/organisms/table/Dav
 import { ChainlinkCTO } from "../../../../../dataAccess/access/cto/ChainlinkCTO";
 import { ChainDecisionTO } from "../../../../../dataAccess/access/to/ChainDecisionTO";
 import { EditActions } from "../../../../../slices/EditSlice";
+import { useAppDispatch } from "../../../../../store";
 import { getChainGotoName } from "../../util/TableUtil";
 
 export const useGetModelChainLinkTableData = (
     selectedChainlinks: ChainlinkCTO[],
     selectedChainDecisions: ChainDecisionTO[],
 ) => {
-    const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
 
-    let bodyData: DavitTableRowData[];
-    bodyData = selectedChainlinks.map((link) => {
+    const bodyData = selectedChainlinks.map((link) => {
         const onClickEdit = () => dispatch(EditActions.setMode.editChainLink(link.chainLink));
         return createLinkColumn(link, selectedChainlinks, selectedChainDecisions, onClickEdit);
     });
